- Updated Food Admin Controller Logic and Added test cases - Modified navbar to include admin dropdown selection

+++ src/main/java/net/ddns/buckeyeflash/controllers/admin/FoodAdminController.java
@@ -6,12 +6,13 @@ import net.ddns.buckeyeflash.models.Food;
 import net.ddns.buckeyeflash.models.datatable.DatatableRequest;
 import net.ddns.buckeyeflash.models.datatable.DatatableResponse;
 import net.ddns.buckeyeflash.repositories.FoodRepository;
-import net.ddns.buckeyeflash.utilities.CommonConstants;
 import net.ddns.buckeyeflash.utilities.PageUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.data.domain.Page;
+import org.springframework.http.MediaType;
+import org.springframework.lang.Nullable;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.web.servletapi.SecurityContextHolderAwareRequestWrapper;
 import org.springframework.stereotype.Controller;
@@ -20,6 +21,7 @@ import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.Optional;
 
 @Controller
 @RequestMapping(value = "/admin/food")
@@ -28,10 +30,12 @@ public class FoodAdminController {
 
     private static final Logger logger = LogManager.getLogger(FoodAdminController.class);
 
-    private static final String FOOD_ATTRIBUTE_NAME = "food";
     private static final String ADD_FOOD_MODAL_TITLE = "Add Food";
     private static final String UPDATE_FOOD_MODAL_TITLE = "Update Food";
     private static final String FOOD_MODAL_CONTENT_FRAGMENT = "fragments/admin/food_fragments :: foodModalContent(title='%s')";
+    private static final String FOOD_ERROR_URL = "/admin/food?status=error";
+    private static final String FOOD_SUCCESS_URL = "/admin/food?status=success";
+    private static final String REDIRECT = "redirect:";
 
     private final FoodRepository foodRepository;
 
@@ -41,8 +45,9 @@ public class FoodAdminController {
 
     @PreAuthorize("hasAnyRole('ADMIN_EDIT','ADMIN_READ')")
     @RequestMapping(method = RequestMethod.GET)
-    public String food(ModelMap modelMap, SecurityContextHolderAwareRequestWrapper securityContextHolderAwareRequestWrapper) {
+    public String food(@RequestParam(required = false) String status, ModelMap modelMap, SecurityContextHolderAwareRequestWrapper securityContextHolderAwareRequestWrapper) {
         modelMap.put("canAdminEdit", securityContextHolderAwareRequestWrapper.isUserInRole("ADMIN_EDIT"));
+        modelMap.put("alertStatus", status);
         logger.info("Admin Food Page Accessed");
         return "pages/admin/food";
     }
@@ -51,9 +56,9 @@ public class FoodAdminController {
     @RequestMapping(value = "/getFoodData", method = RequestMethod.POST)
     public @ResponseBody
     String getFoodData(@RequestBody DatatableRequest request) throws JsonProcessingException {
-        Page<Food> foods = foodRepository.findByTypeStartingWith(request.getSearch().getValue(), PageUtils.getPageRequest(request.getStart(), request.getLength()));
-        ObjectMapper objectMapper = new ObjectMapper();
-        DatatableResponse<Food> datatableResponse = new DatatableResponse<>();
+        final Page<Food> foods = foodRepository.findByTypeStartingWith(request.getSearch().getValue(), PageUtils.getPageRequest(request.getStart(), request.getLength()));
+        final ObjectMapper objectMapper = new ObjectMapper();
+        final DatatableResponse<Food> datatableResponse = new DatatableResponse<>();
         datatableResponse.setDraw(request.getDraw());
         datatableResponse.setRecordsFiltered((int) foods.getTotalElements());
         datatableResponse.setRecordsTotal((int) foods.getTotalElements());
@@ -64,14 +69,16 @@ public class FoodAdminController {
 
     @PreAuthorize("hasRole('ADMIN_EDIT')")
     @RequestMapping(value = "/openFoodModal", method = RequestMethod.POST)
-    public String openInvitationModal(String foodId, ModelMap modelMap) {
+    public String openFoodModal(String foodId, ModelMap modelMap) {
         modelMap.clear();
+        Food food = new Food();
         if (StringUtils.isNotBlank(foodId)) {
-            Food food = foodRepository.findById(Integer.parseInt(foodId)).orElse(new Food());
-            modelMap.addAttribute(FOOD_ATTRIBUTE_NAME, food);
-        } else {
-            modelMap.addAttribute(FOOD_ATTRIBUTE_NAME, new Food());
+            Optional<Food> optionalFood = foodRepository.findById(Integer.parseInt(foodId));
+            if (optionalFood.isPresent()) {
+                food = optionalFood.get();
+            }
         }
+        modelMap.addAttribute("food", food);
         return generateFoodModalContentLocator(foodId);
     }
 
@@ -79,50 +86,46 @@ public class FoodAdminController {
     @RequestMapping(value = "/saveFood", method = RequestMethod.POST)
     public String saveFood(@Valid @ModelAttribute Food food, Errors errors) {
         if (errors.hasErrors()) {
-            return generateFoodModalContentLocator(food.getId());
+            return REDIRECT + FOOD_ERROR_URL;
         }
+        Food tempFood = food;
         if (food.getId() != null) {
-            Food storedFood = foodRepository.findById(food.getId()).get();
-            storedFood.setType(food.getType());
-            storedFood.setDescription(food.getDescription());
-            try {
-                foodRepository.save(storedFood);
-            } catch (Exception e) {
-                logger.error(e.getMessage(), e);
-                return String.format(CommonConstants.MODAL_ERROR_FRAGMENT_TEMPLATE, "Food");
-            }
-        } else {
-            //Add Food
-            try {
-                foodRepository.save(food);
-            } catch (Exception e) {
-                logger.error(e.getMessage(), e);
-                return String.format(CommonConstants.MODAL_ERROR_FRAGMENT_TEMPLATE, "Food");
+            final Optional<Food> optionalFood = foodRepository.findById(food.getId());
+            if (optionalFood.isPresent()) {
+                tempFood = optionalFood.get();
+                tempFood.setType(food.getType());
+                tempFood.setDescription(food.getDescription());
+            } else {
+                return REDIRECT + FOOD_ERROR_URL;
             }
         }
-        return String.format(CommonConstants.MODAL_SUCCESS_FRAGMENT_TEMPLATE, "Food");
+        try {
+            foodRepository.save(tempFood);
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            return REDIRECT + FOOD_ERROR_URL;
+        }
+        return REDIRECT + FOOD_SUCCESS_URL;
     }
 
     @PreAuthorize("hasRole('ADMIN_EDIT')")
-    @RequestMapping(value = "/removeFood", method = RequestMethod.POST)
-    public String removeFood(String foodId) {
+    @RequestMapping(value = "/removeFood", method = RequestMethod.POST, consumes = MediaType.TEXT_PLAIN_VALUE, produces = MediaType.TEXT_PLAIN_VALUE)
+    public @ResponseBody
+    String removeFood(@RequestBody @Nullable String foodId) {
         if (StringUtils.isNotBlank(foodId)) {
-            Food storedFood = foodRepository.findById(Integer.parseInt(foodId)).get();
-            try {
-                foodRepository.delete(storedFood);
-            } catch (Exception e) {
-                logger.error(e.getMessage(), e);
-                return String.format(CommonConstants.MODAL_ERROR_FRAGMENT_TEMPLATE, "Food");
+            final Optional<Food> optionalFood = foodRepository.findById(Integer.parseInt(foodId));
+            if (optionalFood.isPresent()) {
+                final Food storedFood = optionalFood.get();
+                try {
+                    foodRepository.delete(storedFood);
+                    return FOOD_SUCCESS_URL;
+                } catch (Exception e) {
+                    logger.error(e.getMessage(), e);
+                }
             }
-        } else {
-            logger.error("Food Id missing cannot delete");
-            return String.format(CommonConstants.MODAL_ERROR_FRAGMENT_TEMPLATE, "Food");
         }
-        return String.format(CommonConstants.MODAL_SUCCESS_FRAGMENT_TEMPLATE, "Food");
-    }
-
-    private String generateFoodModalContentLocator(Integer foodId) {
-        return generateFoodModalContentLocator(foodId != null ? foodId.toString() : null);
+        logger.error(String.format("Food with id %s cannot be deleted", foodId));
+        return FOOD_ERROR_URL;
     }
 
     private String generateFoodModalContentLocator(String foodId) {
